Time: 67 ms (24.86%), Space: 59.2 MB (74.95%) - LeetHub

+++ 2665-minimum-time-to-repair-cars/2665-minimum-time-to-repair-cars.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
-    bool timeIsSuff(vector<int>& ranks, int cars, long long minGiven) {
+    bool timeIsSuff(vector<int>& ranks, int cars, long long minGiven){
         long long carsDone = 0;
         for (int r : ranks) {
-            long long c2 = minGiven / r;
-            long long c = floor(sqrt(c2));
+            long c2 = minGiven / r;
+            long c = floor(sqrt(c2));
             carsDone += c;
         }
         return carsDone >= cars;
@@ -12,27 +12,30 @@ public:
 
 
 
-//      ...
-//     .   .
-//    .     .
-//   .       .
-//  .         .
-// .           .
-//  .         .
-//   .       .      
-//    .     .
-//     .   .
-//      ...
+    // ______________________________________
+    // |        //      ...                 |
+    // |        //     .   .                |
+    // |        //    .     .               |
+    // |        //   .       .              |
+    // |        //  .         .             |
+    // |        // .   AkShT   .            |
+    // |        //  .  JaIn   .             |
+    // |        //   .       .              |
+    // |        //    .     .               |
+    // |        //     .   .                |
+    // |        //      ...                 |
+    // |____________________________________|
 
 
 
-    long long repairCars(vector<int>& ranks, int cars) {
-        long long l = 1, r = 1e14;
+    long long repairCars(vector<int>& ranks, int cars){
+        long l = 1, r = 1e14;
         while (l < r) {
-            long long mid = (l + r) / 2;
-            if (timeIsSuff(ranks, cars, mid)) {
+            long mid = (l + r) / 2;
+            if(timeIsSuff(ranks, cars, mid)){
                 r = mid;
-            } else {
+            }
+            else{
                 l = mid + 1;
             }
         }
